LOCAL FILES UPDATE:
Remove the dependancy to the online API, and instead extract the ressources from local data files.

Change the files in accordance to it. Instead of having a single 'pokemon_data.json' with online links, we use individual files for each pokemon. Same for the sprites and .png files.

Adding debug functions that were useful to make the project work.