Stats percentiles:
Change the way Pokemon statistics are displayed. Switched to a percentile method to fill the stat bars. This lead to a better visual understanding of the potential of each monster, as we now compare how good he is compared with all Pokemon rather than only the best in the category.
Added flavour text (bad, average, ...) to the stat bars, and a special case if the monster is in the top 5% of a given stat (gives the corresponding percentage).